Multiple alica instances (#241)
* ignore agentAnnouncements of agents using different rolesets

* store masterPlanId in planHash, fix planHash overflow, remove unnecessary check

* use uint64 for planhashes in messages, store planhash in agentQuery as uint64_t, set planhashes in assignment tests

+++ alica_engine/include/engine/containers/AgentQuery.h
@@ -14,6 +14,6 @@ struct AgentQuery
     }
     AgentId senderID;
     uint32_t senderSdk;
-    uint32_t planHash;
+    uint64_t planHash;
 };
 } // namespace alica
